requirements.txt
streamlit
google-genai
transformers
torch
huggingface_hub